refactor CustomTestRunner: use reflections for dynamic test class discovery, update package structure, and enhance execution flow
